requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pydantic-settings==2.2.1
python-dotenv==1.1.0
redis==5.0.1
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
langchain==0.3.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9 
httpx==0.27.2